+++ attributes/default.rb
@@ -8,8 +8,10 @@ default['logstash_stack']['default']['java_opts'] = '-server -Djava.awt.headless
 
 default['logstash_stack']['indexer'] = {}
 
-default['logstash_stack']['web']['java_opts'] = '-server -Djava.awt.headless=true -Xmx256M -Xms64M'
-default['logstash_stack']['web']['domain'] = (fqdn rescue nil)
-
-
 default['logstash_stack']['shipper']['java_opts'] = '-server -Djava.awt.headless=true -Xmx32M -Xms4M'
+
+default['logstash_stack']['kibana3']['version'] = '3.0.0milestone4'
+default['logstash_stack']['kibana3']['url'] = "https://download.elasticsearch.org/kibana/kibana/kibana-#{node['kibana']['kibana3_version']}.tar.gz"
+default['logstash_stack']['kibana3']['checksum'] = '3ebaac69439aa1925c7918e008978b8424840f3bd3910379d4f2bcf5fdfd2118'
+default['logstash_stack']['kibana3']['elasticsearch'] = 'http://127.0.0.1:9002/'
+