+++ metadata.rb
@@ -4,13 +4,13 @@ maintainer_email 'maciej@pasternacki.net'
 license          'MIT'
 description      'Installs/Configures logstash'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.2.0'
+version          '0.2.99'
 
 supports 'debian'
 supports 'ubuntu'
 
 depends 'apache2'
 depends 'apt'
+depends 'ark'
 depends 'java'
-depends 'kibana'
 depends 'runit'