+++ CHANGELOG.md
@@ -3,6 +3,13 @@ CHANGELOG for logstash
 
 This file is used to list changes made in each version of logstash.
 
+0.3.0
+-----
+
+* Bump to logstash 1.2.2 / elasticsearch 0.90.3
+* Bring back packages/ dir
+* [WIP] Kibana 3 proxied by Apache
+
 0.2.0
 -----